pyproject.toml
[tool.poetry]
name = "agave_api"
version = "0.1.6"
description = "A Python client for the Agave API"
authors = ["Joe Petrantoni <79169021+jpetrantoni@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
